requirements.txt
Flask
gunicorn
streamlit

pandas
numpy
seaborn
shap
scikit-learn
joblib
PyYAML

SQLAlchemy
psycopg2-binary
alembic

requests
matplotlib

lightgbm
python-dotenv
